chore: begin ongoing site refinements (initial code cleanup, prep for UI/UX polish)
