+++ src/interfaces/componentProps.types.ts
@@ -1,0 +1,13 @@
+export interface HotAndNewProps {
+  limit: number;
+}
+
+export interface SquareGridContainerProps {
+  category: string;
+}
+
+export interface LineGridProps {
+  category: string;
+  limit?: number;
+  page: number;
+}
