+++ src/interfaces/modal.types.ts
@@ -1,0 +1,4 @@
+export interface ModalState {
+  id: number | null;
+  isOpen: boolean;
+}
