+++ src/interfaces/notification.types.ts
@@ -1,0 +1,10 @@
+export interface Notification {
+  id: number;
+  title: string;
+  message: string;
+  type: "added" | "removed";
+}
+
+export interface NotificationState {
+  notifications: Notification[];
+}
