+++ src/interfaces/product.types.ts
@@ -1,0 +1,8 @@
+export interface ProductCardProps {
+  id: number;
+  title: string;
+  image: string;
+  price: number;
+  discount: number;
+  category?: string;
+}
